folder CGM_workflow remove . CGM_worklfow_main now call worklow placed in pyCGM2/Apps/QtmApps
